add js modules
