+++ webpack.config.js
@@ -2,7 +2,7 @@ const path = require(`path`);
 
 module.exports = {
   mode: `development`,
-  entry: `src/main.js`,
+  entry: `./src/main.js`,
   output: {
     path: path.join(__dirname, `public`),
     filename: `bundle.js`
